feat: 비동기 전역상태 처리 및 구조 정리
