+++ README.md
@@ -35,3 +35,4 @@
 - [x] 값에 의미에 적합하도록 이름 수정
 - [x] issue 데이터에서 컴포넌트에 필요한 값만 골라서 가져오는 함수 생성
 - [x] 리스트 가져오는 파이프 수정
+- [x] 비동기 전역상태 처리
